Avoid ASAN complaining about signed-unsigned conversions

+++ src/Vec.h
@@ -245,7 +245,7 @@ void vec<T>::capacity(int32_t min_cap)
     }
 
     // NOTE: grow by approximately 3/2
-    uint32_t add = imax((min_cap - cap + 1) & ~1, ((cap >> 1) + 2) & ~1);
+    uint32_t add = imax((min_cap - (int32_t)cap + 1) & ~1, (((int32_t)cap >> 1) + 2) & ~1);
     if (add > std::numeric_limits<uint32_t>::max() - cap
         || (((data = (T*)::realloc(data, (cap += (uint32_t)add) * sizeof(T))) == NULL)
             && errno == ENOMEM)
